Author: Christos Tsantilas <chtsanti@users.sourceforge.net> Bug 2219: pconn status logic errors when getting chunks
The problem here is that chunk-decoding was wrongly using the eof to mark
the last chunk (look at method HttpStateData::decodeAndWriteReplyBody).
The eof used in HttpStateData to mark that the connection probably closed so
another flag must be used to mark that the last chunk received.

This patch uses the HttpStateData::lastChunk flag to mark that the last chunk
received. Maybe the flag should be moved to  http_state_flags struct or a
general flag "complete" should used to mark that the whole body received.