Abort sendMoreData if our socket is being closed to avoid comm.cc:2032: "!fd_table[fd].closing()" assertions in comm_write.
The assert happens when sendMoreData is triggered by an event on the other
(server) side.  The server side does not know that client side (or something
else) have decided to close the client socket. The close callback has not been
dialed yet so the client state has not been invalidated and non-async from
Store callbacks can reach it.

A better fix for this would be to convert more store callbacks to AsyncCalls,
but that may have to wait for client_side rewrite.

Also assert that USE_ZPH_QOS code does not use a negative socket descriptor.
The "fd = conn != NULL ? conn->fd : -1" code seems to imply that our
descriptor may be negative. I do not know whether that can actually happen.


This fix was a part of the recent cleanup effort but got lost when I split
the changes into several MERGE requests :-(.