Bug 740: allow external acl's to use reply headers in format
Adds a small bit of token syntax to external_acl_type format.

  %>{Header}	HTTP request header
  %>{Hdr:member}
  		HTTP request header list member
  %>{Hdr:;member}
		HTTP request header list member using ; as
  		list separator. ; can be any non-alphanumeric
		character.

  %<{Header}	HTTP reply header
  %<{Hdr:member}
  		HTTP reply header list member
  %<{Hdr:;member}
  		HTTP reply header list member using ; as
  		list separator. ; can be any non-alphanumeric
		character.

Basically the < and > are new following the existing meaning of their 
direction in other tokens to match request/reply.

Old format of %{} is left as request header but with WARNING (1) level 
noise at configure time indicating the new syntax.