Bug #2447: Segfault on failed TCP DNS query
This patch makes Squid log the DNS failure if it fails to connect to
a DNS server over TCP. TCP is required if the DNS response do not fit
within the small DNS UDP packet size (ca 0.5kb).

This patch apparently do not solve the segmentation fault, but at least
logs the condition properly.