SourceFormat: Convert md5checker script into a recursive formatter+validator
This converts the md5 validation script from a single layer to a
recursive script.

Given a directory path calls the formater.pl script for all .h .c .cc and
.cci files within. Validating each file conversion as it goes and aborts
at the first error found.

This is intended for maintenance of the central Squid-3 repository code.

NP: As with the original checker its a bash script.
    Probably non-portable right now without adaptions.