Bug 2393: DNS requests getting stuck in idns queue
Port of Squid-2 fix by Henrik Nordstrom

There seems to be two problems here.

a) On TCP communication failure Squid starts to restransmit the TCP query as
fast as it can, with no bounds checking. Keeps doing that until there is a
response or Squid is restarted...

b) For some reason the retransmit queue seems to halt for you. I have not yet
reproduced this issue, but may be related to the first..