Update public IPv6 configure sub-options.
Drop all mention of CNAME recursion. Still buggy and experimental.

Alter --with-ipv4-mapping to default ON.
This is harmless if system defaults it to ON. But resolves problems
seen with systems where OS default is OFF disabling this advanced
hybrid-stack feature which squid needs.
Seen in some Linux, FreeBSD 7.1, OpenBSD, and Vista.