Fix assertion in DskIO and Ufs unit tests
Classes whose initialization is implied from the presence of a soliton
do not get initialized when their soliton is hidden away in a library.
Only when the object is linked to a main app.

DiskIO modules are such and require a hook to probe the soliton before
initialization kicks off. This patch adds such a hook function which is
called from the manager setup function to probe all built modules and
ensure their existance and registration is done before initializing.

Also adds a lot of function documentation in various places.