Author: Regardt van de Vyver <squid@vdvyver.net> Bug 2555: Fixes to SNMP-MIB
* Modified imports as Integer32 and TEXTUAL-CONVENTION are external defines to
this MIB, it's a compliance issue for validation

* Updated the order of Revisions as the MIB spec requires them to be in reverse
choronological order

* moved the nlanr define to a valid location as it may not be the first defined
item in a module (the MODULE IDENTITY must be)

* Modified various "ACCESS" statements to "MAX-ACCESS" due to changes in the
MIB spec requiring the change of defines.

* Modified various interative table entries to "not-accessible". This is due to
the fact that the root element should not be directly accessed, rather the
generated table entries matching these values. This is not strictly neccesary
but does make the file validate on severity 3 instead of severity 2.

* Modified cacheMedianTime to have a valid range. Ranges are REQUIRED for any
object used as an INDEX

* Modified cachePeerPortHTTP and cachePeerPortIcp to use a new type
"ValidPort". This is to meet a requirement that we cannot define valid port
ranges in a sequence defenition.

* Created a new define ValidPort so that we can specify a range of valid values
for ports

* In some object defines STATUS was set to mandatory, this is however not a
valid MIB STATUS value and has thus been fixed to current. These items
specifically caused the MIBs to fail compile completely.

* cacheClientEntry had a duplicate oid and after checking the source code I
updated its oid to the correct value.



SIDE NOTE:
Squid-2 and 3.0 have a different structure for cachePeerEnty enum. This may
cause some confusion if the squid3 mib is used with older squid releases.