Author: Christos Tsantilas <chtsanti@users.sourceforge.net> Fix assertion in comm_write from idnsSentQueryVC
The idnsSentQueryVC function called as AsyncCall. There is the possibility
when this function called the fd has start closing but not realy closed yet.
In this case this function will try to do a comm_write (idnsDoSendQueryVC
function) on sockect which closing and an assertion will triggered.
An extra test needed here to test if the socket closing and if yes just return.