TestBed: add slave tester script
This is draft #2 of the multi-machine script. I have it running locally.

As an intermediary step to kick wider testing until we get something
better going.  This script can be run by cron and sends an email to
noc@squid-cache.org detailing the local system OS, CPU, and build results.

Requires:
    sendmail or mailer providing sendmail API
    bzr with checkout of (trunk, SQUID_3_1, or SQUID_3_0)

TODO:
    check that the binaries called are portable