Bug 2601: Hack. Convert IPv4 netmasks to CIDR in IPv6-enabled mode
se bug 2601 for trace demonstrating the effect of masking an IPv6 address
with and IPv4 netmask intead of a CIDR mask.

This hack, locates what CIDR mask was _probably_ meant to be in its
native protocol format. Then resets the mask to that CIDR form.

This will completely crap out with a security fail-open if the admin is
playing mask tricks.  However, thats their fault, and we do warn loudly.