SourceLayout: setup libcompat.la for portability primitives
This library forms an underlayer which intends to be seamless for the rest
of Squid and the various OS on which it runs.

For code to be eligible for inclusion at this lowest layer it must be
emulating or wrapping a piece of OS-provided API for use on other OS 
where it is not provided.

Shared code which is unique to Squid is not portability.

API mappings which are OS-specific are all defined for their particular
OS in the os/* files.

API mappings which are provided by multiple OS or support libraries
are mapped in compat/*

Emulators are acceptable, though the aim should be to inline or template
most of the code so only .h are really needed.