Complete Interception multiple NAT support
Bug 2618: fix ipfilter transparent proxy
Thanks to John Wehle <john@feith.com> for providing the code fixes.

Also:
  - cleans up some obscurity over data source and sink for me/client IPs.
  - cleans up existing debugs
  - adds new debugs to show NAT inputs and results at level-5
    (non-result is common and left at level-9)
  - adds new debugs to show TPROXY result at level-5 like NAT results
  - move IPF lookup into it's own function
  - move PF interception into its own function

With this all of the transparent build options are independent, and may be
used in any combination. Squid is no longer bound to the single-firewall 
interception support.

NP: one small note; the PF lookup is slightly weird due to its altering
the local client address from the NAT information.
Not sure if this is needed, leaving it alone for now.