Auto-Docs: Add \cpptest directive
This is a custom directive just for squid.
It's used by adding to the declaration documentation of a function or
class method and listing the name of a unit-test function which
tests the documented function/method. It may be used multiple times
in a single function if there are multiple unit-tests for it.
Though in general multiple unit-tests sequences should be done in one
testing function.