Collate and HTMLize the current Squid manual pages
This is expected to create a *-manuals.tar.gz bundle like the cfgman
bundle before it with raw and HTML versions of all man files currently
in the Squid sources.  For the purpose of providing web page versions
as authorative reference documentation.

This is one of those things that really has to be tested live on snapshot
generation. I'm only sure that groff exists in the spots its needed and
that the creation of the tar.gz should work. More work and testing will
likely be needed before public pages are available.

TODO:
    Test this alteration, and convert the bundle to live web pages.
    Add man .8 pages for many of the helpers.