Correct TestBed make checks.
distcheck drops any custom configure options. We can't use it
outside the 00-default test.

We can keep check, and should add any others that may be useful if
they can be identified and shown not to re-run configure.

Also move some old distclean hacks that may be obsolete now out of
the way to a rare occurance when thay might still be relevant.