Author: Alin Nastac <mrness@gentoo.org> Cleanup: deprecate ugly hack for sys/capability.h
Needed once to prevent build clashes between libc and sys/capabilility.h
headers. As of libcap2.16 that is no longer the case and in fact prevents
a clean build.

Left in libcompat but removed from active build until its clear whether
full erasure is possible or if some systems still need it.