Cleanup: Split some enums into their own files. auto-generate their text name arrays.
Also updates the enum strings generator to work with any given header file.
The header does however need to be named identical to the enum and not pull
in any other definitions that require .cc content. The generated .cc will
only contain the minimal includes and enum string array global.