Author: Tim Starling <unknown> Port 2.7: Unique sequence number per log line
Create and enable logging of a unique sequence number per log entry.

Adjusted to 64-bit to handle long Squid up times with very large logged
line counts.

NP: not quite sure about the incrementation. Write appears to be called
once per log entry but this is not tested under enough conditions to be
certain yet. The old lineEnd call from Squid-2 is not present in Squid-3.